requirements.txt
djangorestframework==3.14.0
Django==4.1
djangorestframework-simplejwt==5.2.2